feat: Add reapply functionality for organizations
This commit adds the ability to reapply for organizations in the `fetchOrgs.php` file. It introduces a new case in the switch statement for the "Reapply" action, which updates the status of the organization to "Pending" in the database. The `Reapply` method is also added to the `DB` class in the `index1.php` file, which handles the database connection and executes the update query. Additionally, the `Reapply` method is implemented in the `Organizations` class in the `organizations.php` file, which performs the actual update query.

The `org_profile.html` file is also modified to include a new button for reapplying. When clicked, it sends a request to the `fetchOrgs.php` file to trigger the reapply action.

This feature allows organizations to easily reapply for their status to be reconsidered.